Refactor: Move types and queries to new directory structure
- Deleted old type definitions for shows and utilities.
- Created new type definitions for shows and utilities in the _types directory.
- Updated imports in components to reflect new paths for queries and types.
- Added new queries for feature flags and user favorites management.
- Implemented caching for API requests in the new queries structure.
