+++ app/favorites/page.tsx
@@ -4,10 +4,10 @@ import { LocalSearch } from '@/components/utils/SearchComponent';
 import { useDbPropsCtx } from '../_context/DbProps';
 import LandingCard from '@/components/landing/LandingMovieCard';
 import { SearchXIcon } from 'lucide-react';
-import { Movie } from '../types/movies';
+import { Movie } from '../_types/movies';
 import { movie } from '@prisma/client';
 import { useMainCtx } from '../_context/Main';
-import { ListType } from '../types/utils';
+import { ListType } from '../_types/utils';
 import { SelectComponent } from '@/components/utils/Select';
 import { filterType } from '@/lib/arrys';
 import PageTitle from '@/components/utils/texts/PageTitle';
