+++ components/general/LandingComponentClient.tsx
@@ -2,10 +2,10 @@
 import React from 'react';
 import HeroMovieImg from '../landing/HeroMovieImage';
 import HeroMovieInfo from '../landing/HeroMovieInfo';
-import { Movie } from '@/app/types/movies';
-import { Show } from '@/app/types/show';
+import { Movie } from '@/app/_types/movies';
+import { Show } from '@/app/_types/show';
 import useHero from '@/app/_hooks/useHero';
-import { ListType } from '@/app/types/utils';
+import { ListType } from '@/app/_types/utils';
 
 const LandingComponentClient = ({ type, items }: { type: ListType; items: Movie[] | Show[] }) => {
   const { movie, time, direction, intervalTime } = useHero(items);
