+++ components/landing/LandingMovieCard.tsx
@@ -1,12 +1,12 @@
-import { Movie } from '@/app/types/movies';
+import { Movie } from '@/app/_types/movies';
 import Link from 'next/link';
 import React from 'react';
 import { imageCardUrl } from '@/lib/utils';
 import ImageComponent from '../utils/ImageComponent';
 import CarouselItem from '../carousel/CarouselItem';
 import FavAndReview from '../utils/buttons/FavAndReview';
-import { Show } from '@/app/types/show';
-import { ListType } from '@/app/types/utils';
+import { Show } from '@/app/_types/show';
+import { ListType } from '@/app/_types/utils';
 
 const LandingCard = ({ type, movie }: { type: ListType; movie: Movie | Show }) => {
   const title = movie ? (movie as Movie).title || (movie as Show).original_name : '';
