+++ components/utils/buttons/FavAndReview.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 import Ratings from '../texts/Ratings';
 import Favorite from './Favorite';
 import { movie } from '@prisma/client';
-import { Movie } from '@/app/types/movies';
-import { Show } from '@/app/types/show';
-import { ListType } from '@/app/types/utils';
+import { Movie } from '@/app/_types/movies';
+import { Show } from '@/app/_types/show';
+import { ListType } from '@/app/_types/utils';
 
 const FavAndReview = ({ type, movie, title }: { type: ListType; movie: Movie | Show; title: string }) => {
   return (
